refactor(shared): apply layered structure to packages/shared/src/
Move 24 source/test files into domain/, plus 3 into geo/. Collapse three
root-level barrel files (errors.ts, types.ts, api.ts) into their
dir-index counterparts. Public surface (@fyndstigen/shared exports)
unchanged; only internal paths and edge-function subpaths shift.

Subdirs:
- domain/ — booking, booking-lifecycle, booking-service, block-sale,
            business-import, market-mutation, route-mutation,
            route-optimizer, opening-hours (+ co-located tests),
            booking-logic.test, payment-amounts.test
- geo/    — index (was geo.ts), routing (+ tests)

Root barrels collapsed:
- errors.ts content → errors/index.ts (next to messages.sv.ts)
- api.ts content    → api/index.ts (next to bookings.ts, flea-markets.ts)
- types.ts shim     → deleted (types/index.ts already canonical)

Edge-function imports updated (6 files, ~12 sites):
- @fyndstigen/shared/booking → domain/booking
- @fyndstigen/shared/booking-lifecycle → domain/booking-lifecycle
- @fyndstigen/shared/block-sale → domain/block-sale
- @fyndstigen/shared/business-import → domain/business-import

Allow-list (scripts/check-edge-imports.mjs) updated to match. Pre-existing
miss for contracts/route-create-anon and types/index added.

packages/shared/package.json exports map adds ./domain/*, ./geo/*,
./types/* glob entries. web/vitest.config.ts alias updated.

Files staying at root: index.ts (public barrel), crypto.ts (+ test),
format.ts, stripe-webhook.ts (+ test), deps.ts, deps-factory.ts (+ test),
errors.test.ts, payment-flow.test.ts (integration test), adapters.test.ts.

Closes #89
