+++ packages/shared/src/domain/booking-lifecycle.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 import { applyBookingEvent, type BookingEvent, type BookingPatch } from './booking-lifecycle'
-import type { Booking, BookingStatus } from './types'
+import type { Booking, BookingStatus } from '../types'
 
 const FIXED_NOW = new Date('2026-04-21T12:00:00.000Z')
