+++ packages/shared/src/domain/booking-logic.test.ts
