+++ packages/shared/src/domain/booking.test.ts
