+++ packages/shared/src/domain/opening-hours.test.ts
