+++ packages/shared/src/domain/payment-amounts.test.ts
