+++ packages/shared/src/domain/route-optimizer.test.ts
