+++ packages/shared/src/domain/route-optimizer.ts
@@ -1,4 +1,4 @@
-import type { Coord } from './types/domain'
+import type { Coord } from '../types/domain'
 
 export type Stop = Coord & {
   id: string
