--- packages/shared/src/types.ts
@@ -1,4 +1,0 @@
-// This file is kept for import-path compatibility.
-// All type definitions have moved to types/index.ts.
-// Do not add new types here — add them in types/domain.ts or types/db.ts.
-export * from './types/index'
