+++ supabase/functions/stripe-payment-cancel/index.ts
@@ -5,7 +5,7 @@ import { appError } from '@fyndstigen/shared/errors.ts'
 import { createSupabaseBookingRepo } from '@fyndstigen/shared/adapters/supabase/booking-repo.ts'
 import { createStripeBookingGateway } from '@fyndstigen/shared/adapters/stripe/booking-stripe-gateway.ts'
 import { StripePaymentCancelInput, StripePaymentCancelOutput } from '@fyndstigen/shared/contracts/stripe-payment-cancel.ts'
-import type { BookingEvent } from '@fyndstigen/shared/booking-lifecycle.ts'
+import type { BookingEvent } from '@fyndstigen/shared/domain/booking-lifecycle.ts'
 
 defineEndpoint({
   name: 'stripe-payment-cancel',
