+++ web/vitest.config.ts
@@ -16,7 +16,9 @@ export default defineConfig({
       // vitest resolve the in-tree source so worktree-local changes are visible
       // to tests. Harmless on main — the alias still points at the same path.
       '@fyndstigen/shared/deps-factory': path.resolve(__dirname, '../packages/shared/src/deps-factory.ts'),
-      '@fyndstigen/shared/block-sale': path.resolve(__dirname, '../packages/shared/src/block-sale.ts'),
+      // domain/* and geo/* subpath aliases — resolves to worktree source, bypassing the
+      // hoisted node_modules symlink that still points at main.
+      '@fyndstigen/shared/domain/block-sale': path.resolve(__dirname, '../packages/shared/src/domain/block-sale.ts'),
       '@fyndstigen/shared': path.resolve(__dirname, '../packages/shared/src/index.ts'),
     },
   },
